requirements.txt
asttokens==2.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cssselect==1.2.0
decorator==5.1.1
executing==1.2.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.0
idna==3.4
ipython==8.6.0
jedi==0.18.1
jieba3k==0.35.1
joblib==1.2.0
lxml==4.9.1
matplotlib-inline==0.1.6
newspaper3k==0.2.8
nltk==3.7
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.32
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.6.1
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.64.1
traitlets==5.5.0
urllib3==1.26.12
wcwidth==0.2.5
